// src/app/shared/models/offer.config.model.ts
export class OfferConfiguration {
    
    idOferta: string;
    idEstatus: number;
    inicio: string;
    fin: string;
    fechaModificacion: string;
    idRecordatorioInicio: string;
    idRecordatorioFin: string;
    idTipoAlerta: Array<number>;
    distancia: number;
    precio: number;
    cupon: string;
    idTemperatura: number;
    idSexo: number;
    edadDe: number;
    edadA: number;

}